Finished

+++ MauiCrudJuht/FrontPage.xaml.cs
@@ -4,6 +4,7 @@ namespace MauiCrudJuht
     {
         private readonly LocalDbService _dbService;
         private int _editCustomerId;
+        private bool _isEditing;
 
         public FrontPage(LocalDbService dbService)
         {
@@ -37,6 +38,7 @@ namespace MauiCrudJuht
                 });
 
                 _editCustomerId = 0;
+                _isEditing = false;
             }
 
             nameEntryField.Text = string.Empty;
@@ -51,7 +53,7 @@ namespace MauiCrudJuht
         private async void ListView_ItemTapped(object sender, ItemTappedEventArgs e)
         {
             var customer = (Customer)e.Item;
-            var action = await DisplayActionSheet("Action", "Cancel", null, "Edit", "Delete");
+            var action = await DisplayActionSheet("Action", "Cancel", null, "Edit", _isEditing ? null : "Delete");
 
             switch (action)
             {
@@ -61,12 +63,15 @@ namespace MauiCrudJuht
                     nameEntryField.Text = customer.CustomerName;
                     emailEntryField.Text = customer.Email;
                     mobileEntryField.Text = customer.Mobile;
+                    _isEditing = true;
 
                     break;
                 case "Delete":
-
-                    await _dbService.Delete(customer);
-                    ListView.ItemsSource = await _dbService.GetCustomers();
+                    if (!_isEditing)
+                    {
+                        await _dbService.Delete(customer);
+                        ListView.ItemsSource = await _dbService.GetCustomers();
+                    }
                     break;
             }
